More proxy improvements (#20)
* log requests within CONNECT tunnel

* set `Connection: close`

goproxy doesn't hang up the upstream connection when the downstream
connection closes. So, it'll read in the next request and then fail
to send it downstream. In this case it 502's. It would be nicer if
it could close the upstream connection preemtively so the cient
could re-dial. Elixir's HTTPoison tries pooling requests over the
CONNECT tunnel and this results in 502's bubbling up to the calling
app. Setting `Connection: close` prevents this pooling.